+++ src/interpreter/LoxFunction.hpp
@@ -10,6 +10,6 @@ class LoxFunction : public LoxCallable {
 public:
   LoxFunction(std::shared_ptr<Function> declaration, std::shared_ptr<Environment> closure) : declaration{ declaration }, closure{ closure } {};
   std::any call(Interpreter &interpreter, std::vector<std::any> arguments) override;
-  int arity() override;
+  size_t arity() override;
   friend std::ostream& operator<<(std::ostream& out, const LoxFunction& function);
 };
